+++ src/activities/home/HomeActivity.h
@@ -75,6 +75,7 @@ class HomeActivity final : public Activity {
   void onFileTransferOpen();
   void onOpdsBrowserOpen();
   void onReadingStatsOpen();
+  void openReadingStatsDirect();
   void onSavedItemsOpen();
 
   int getMenuItemCount() const;